requirements.txt
# Functional
dash
dash-bootstrap-components
pandas
plotly
requests
SQLAlchemy
python-dateutil

# Linting
black
mypy
flake8
flake8-bugbear
tox